backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
python-dotenv==1.2.2
google-adk==1.27.1
google-genai==1.67.0
playwright==1.58.0
browserbase==1.6.0
google-cloud-storage==3.9.0
google-cloud-firestore==2.21.0
pydantic==2.12.5
websockets==15.0.1
